+++ packages/manager/apps/cloud/client/app/dbaas/logs/detail/streams/add/logs-streams-add.controller.js
@@ -3,7 +3,9 @@ class LogsStreamsAddCtrl {
     $q,
     $state,
     $stateParams,
+    $translate,
     LogsStreamsService,
+    LogsTokensService,
     CucControllerHelper,
     CucCloudMessage,
     LogsConstants,
@@ -11,8 +13,10 @@ class LogsStreamsAddCtrl {
     this.$q = $q;
     this.$state = $state;
     this.$stateParams = $stateParams;
+    this.$translate = $translate;
     this.serviceName = this.$stateParams.serviceName;
     this.LogsStreamsService = LogsStreamsService;
+    this.LogsTokensService = LogsTokensService;
     this.CucControllerHelper = CucControllerHelper;
     this.CucCloudMessage = CucCloudMessage;
     this.LogsConstants = LogsConstants;
@@ -21,7 +25,12 @@ class LogsStreamsAddCtrl {
     this.storageDurations = this.LogsStreamsService.getStorageDurations();
     this.storageTargets = this.LogsStreamsService.getStorageTargets();
     this.storageContents = this.LogsStreamsService.getStorageContents();
-    this.coldStoragePrice = { price: '' };
+    this.coldStoragePrice = { PCS: { price: '' }, PCA: { price: '' } };
+    this.indexingStoragePrice = {
+      FirstStep: { price: '' },
+      SecondStep: { price: '' },
+    };
+    this.availableRetentions = [];
     this.initLoaders();
   }
 
@@ -31,12 +40,10 @@ class LogsStreamsAddCtrl {
    * @memberof LogsStreamsHomeCtrl
    */
   initLoaders() {
-    this.options = this.CucControllerHelper.request.getArrayLoader({
+    this.defaultCluster = this.CucControllerHelper.request.getHashLoader({
       loaderFunction: () =>
-        this.LogsStreamsService.getSubscribedOptions(this.serviceName),
+        this.LogsTokensService.getDefaultCluster(this.serviceName),
     });
-    this.options.load();
-
     this.mainOffer = this.CucControllerHelper.request.getArrayLoader({
       loaderFunction: () =>
         this.LogsStreamsService.getMainOffer(this.serviceName),
@@ -49,26 +56,69 @@ class LogsStreamsAddCtrl {
       loaderFunction: () =>
         this.LogsStreamsService.getAccountDetails(this.serviceName),
     });
-
-    this.accountDetails.load().then(() => {
-      this.ovhSubsidiary = this.accountDetails.data.me.ovhSubsidiary;
-      this.$q.all([this.mainOffer.load(), this.catalog.load()]).then(() => {
-        if (
-          this.mainOffer.data.planCode === this.LogsConstants.basicOffer &&
-          !this.isEdit
-        ) {
-          this.stream.data.webSocketEnabled = false;
-        }
+    this.accountDetails
+      .load()
+      .then(() => {
+        this.ovhSubsidiary = this.accountDetails.data.me.ovhSubsidiary;
+        return this.$q.all([this.mainOffer.load(), this.catalog.load()]);
+      })
+      .then(() => {
         const selectedCatalog = this.catalog.data.plans.find(
           (plan) => plan.planCode === this.mainOffer.data.planCode,
         );
-        const coldstorage = selectedCatalog.addonsFamily.find(
-          (addon) => addon.family === this.LogsConstants.COLDSTORAGE,
+        const selectedFamily = selectedCatalog.addonsFamily.find(
+          (addon) => addon.family === this.LogsConstants.ADD_ON_FAMILY.NEW,
         );
-        this.coldStoragePrice.price =
-          coldstorage.addons[0].plan.details.pricings.default[0].price.text;
+        const indexingCapacities = selectedFamily.addons.find(
+          (add) =>
+            add.plan.planCode ===
+            this.LogsConstants.CONSUMPTION_REFERENCE.STREAM,
+        );
+        const coldstoragePCACapacities = selectedFamily.addons.find(
+          (add) =>
+            add.plan.planCode ===
+            this.LogsConstants.CONSUMPTION_REFERENCE.COLDSTORAGE_PCA,
+        );
+        const coldstoragePCSCapacities = selectedFamily.addons.find(
+          (add) =>
+            add.plan.planCode ===
+            this.LogsConstants.CONSUMPTION_REFERENCE.COLDSTORAGE_PCS,
+        );
+        const indexingFirstStepPrice = indexingCapacities.plan.details.pricings.default.find(
+          (capabilities) =>
+            capabilities.capacities.includes(
+              this.LogsConstants.CONSUMPTION_CAPACITY,
+            ) &&
+            capabilities.maximumQuantity ===
+              this.LogsConstants.INDEXING_TIERING,
+        );
+        const indexingSecondStepPrice = indexingCapacities.plan.details.pricings.default.find(
+          (capabilities) =>
+            capabilities.capacities.includes(
+              this.LogsConstants.CONSUMPTION_CAPACITY,
+            ) &&
+            capabilities.minimumQuantity ===
+              this.LogsConstants.INDEXING_TIERING + 1,
+        );
+        const coldstoragePCA = coldstoragePCACapacities.plan.details.pricings.default.find(
+          (capabilities) =>
+            capabilities.capacities.includes(
+              this.LogsConstants.CONSUMPTION_CAPACITY,
+            ),
+        );
+        const coldstoragePCS = coldstoragePCSCapacities.plan.details.pricings.default.find(
+          (capabilities) =>
+            capabilities.capacities.includes(
+              this.LogsConstants.CONSUMPTION_CAPACITY,
+            ),
+        );
+        this.coldStoragePrice.PCA.price = coldstoragePCA.price.text;
+        this.coldStoragePrice.PCS.price = coldstoragePCS.price.text;
+        this.indexingStoragePrice.FirstStep.price =
+          indexingFirstStepPrice.price.text;
+        this.indexingStoragePrice.SecondStep.price =
+          indexingSecondStepPrice.price.text;
       });
-    });
 
     if (this.$stateParams.streamId) {
       this.isEdit = true;
@@ -83,6 +133,30 @@ class LogsStreamsAddCtrl {
     } else {
       this.isEdit = false;
       this.stream = this.LogsStreamsService.getNewStream();
+
+      this.defaultCluster.load().then((cluster) => {
+        this.availableRetentions = cluster.retentions.reduce(
+          (retentionsList, retention) => {
+            if (retention.isSupported) {
+              const updatedRetention = retention;
+              if (updatedRetention.duration) {
+                updatedRetention.label = moment
+                  .duration(updatedRetention.duration)
+                  .humanize();
+              } else {
+                updatedRetention.label = this.$translate.instant(
+                  'streams_disk_full',
+                );
+              }
+              retentionsList.push(updatedRetention);
+            }
+            return retentionsList;
+          },
+          [],
+        );
+        this.stream.data.retentionId = cluster.defaultRetentionId;
+      });
+      this.stream.data.retentionId = this.defaultCluster.defaultRetentionId;
     }
   }
 
@@ -92,6 +166,52 @@ class LogsStreamsAddCtrl {
     } else {
       this.createStream();
     }
+  }
+
+  getColdStoragePrice() {
+    return this.$translate.instant(
+      'streams_cold_storage_price',
+      {
+        t0: this.coldStoragePrice[this.stream.data.coldStorageTarget].price,
+        t1: this.LogsConstants.COLDSTORAGE_INCREMENT,
+      },
+      undefined,
+      false,
+      'sceParameters', // Expose devise symbol from API without sanitization
+    );
+  }
+
+  getIndexingPrices() {
+    return this.$translate.instant(
+      'logs_streams_enable_indexing_description',
+      {
+        t0: this.indexingStoragePrice.FirstStep.price,
+        t1: this.LogsConstants.INDEXING_TIERING,
+        t2: this.indexingStoragePrice.SecondStep.price,
+      },
+      undefined,
+      false,
+      'sceParameters', // Expose devise symbol from API without sanitization
+    );
+  }
+
+  getIndexingMaxSizeText() {
+    if (this.stream.data.indexingMaxSize === 0) {
+      return this.$translate.instant('logs_streams_no_limit');
+    }
+    return '';
+  }
+
+  getIndexingNotificationText() {
+    return this.$translate.instant('logs_streams_alert_notification_detail', {
+      t0: this.stream.data.indexingMaxSize,
+    });
+  }
+
+  getIndexingPauseText() {
+    return this.$translate.instant('logs_streams_pause_indexing_detail', {
+      t0: this.stream.data.indexingMaxSize,
+    });
   }
 
   /**