+++ packages/manager/apps/web/client/app/hosting/hosting.routes.js
@@ -23,6 +23,7 @@ export default /* @ngInject */ ($stateProvider) => {
           emailOptionServiceInfos.map(({ serviceId }) =>
             ovhManagerProductOffersDetachService
               .getAvailableDetachPlancodes(serviceId)
+              .catch(() => [])
               .then((plancodes) => ({
                 serviceId,
                 detachPlancodes: plancodes,
@@ -41,21 +42,62 @@ export default /* @ngInject */ ($stateProvider) => {
             emailOptionIds.map((emailOptionId) =>
               hostingEmailService
                 .getEmailOptionServiceInformation(serviceName, emailOptionId)
-                .then(
-                  ({ resource }) =>
-                    OvhApiEmailDomain.v6().serviceInfos({
+                .then(({ resource }) =>
+                  OvhApiEmailDomain.v6()
+                    .serviceInfos({
                       serviceName: resource.name,
-                    }).$promise,
-                ),
+                    })
+                    .$promise.catch(() => null),
+                )
+                .catch(() => null),
             ),
           )
-          .then((servicesInformation) => servicesInformation.flatten()),
+          .then((servicesInformation) =>
+            servicesInformation
+              .filter((information) => information !== null)
+              .flatten(),
+          ),
       pendingTasks: /* @ngInject */ (HostingTask, serviceName) =>
         HostingTask.getPending(serviceName).catch(() => []),
+      privateDatabasesIds: /* @ngInject */ (HostingDatabase, serviceName) =>
+        HostingDatabase.getPrivateDatabaseIds(serviceName).catch(() => []),
+      privateDatabasesDetachable: /* @ngInject */ (
+        $q,
+        ovhManagerProductOffersDetachService,
+        PrivateDatabase,
+        privateDatabasesIds,
+      ) =>
+        $q
+          .all(
+            privateDatabasesIds.map((id) =>
+              PrivateDatabase.getServiceInfos(id).catch(() => null),
+            ),
+          )
+          .then((privateDatabasesInformation) =>
+            privateDatabasesInformation
+              .filter((information) => information !== null)
+              .flatten(),
+          )
+          .then((privateDatabasesInformation) =>
+            $q.all(
+              privateDatabasesInformation.map(({ domain, serviceId }) =>
+                ovhManagerProductOffersDetachService
+                  .getAvailableDetachPlancodes(serviceId)
+                  .catch(() => [])
+                  .then((plancodes) => ({
+                    optionId: domain,
+                    serviceId,
+                    detachPlancodes: plancodes,
+                  })),
+              ),
+            ),
+          ),
       serviceName: /* @ngInject */ ($transition$) =>
         $transition$.params().productId,
       goToDetachEmail: /* @ngInject */ ($state) => () =>
         $state.go('app.hosting.detachEmail'),
+      goToDetachPrivateDB: /* @ngInject */ ($state) => () =>
+        $state.go('app.hosting.database.detachPrivate'),
       goToHosting: /* @ngInject */ ($state, $timeout, Alerter) => (
         message = false,
         type = 'success',