+++ packages/manager/modules/product-offers/CHANGELOG.md
@@ -1,3 +1,21 @@
+# [1.2.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-product-offers@1.1.1...@ovh-ux/manager-product-offers@1.2.0) (2020-05-13)
+
+
+### Features
+
+* **product-offers:** force camel case payment method param ([38917ff](https://github.com/ovh/manager/commit/38917ff88e0c56ea41badf7c233fa22b4c6b3975))
+
+
+
+## [1.1.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-product-offers@1.1.0...@ovh-ux/manager-product-offers@1.1.1) (2020-05-04)
+
+
+### Bug Fixes
+
+* **workflow:** autopay order only if customer has payment mean ([#2881](https://github.com/ovh/manager/issues/2881)) ([5b73b3c](https://github.com/ovh/manager/commit/5b73b3c5cf4ef7388787524ea47ec7a3eff52290))
+
+
+
 # [1.1.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-product-offers@1.0.4...@ovh-ux/manager-product-offers@1.1.0) (2020-04-28)